LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/path-objects/2d.path.bezierCurveTo.nonfinite-expected.txt

2d.path.bezierCurveTo.nonfinite

bezierCurveTo() with Infinity/NaN is ignored


PASS bezierCurveTo() with Infinity/NaN is ignored
